% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEPEst-package.R
\docType{package}
\name{DEEPEst-package}
\alias{DEEPEst-package}
\alias{DEEPEst}
\title{The 'DEEPEst' package.}
\description{
Hierarchical Bayesian Estimation for DEEP with Stan
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
