% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_plot_cor.R
\name{Risk_plot_cor}
\alias{Risk_plot_cor}
\title{Plot Correlation between Estimates for Risk Preferences}
\usage{
Risk_plot_cor(project_name, num_question_Est, type_theta, path, parameter1,
  parameter2)
}
\arguments{
\item{project_name}{The name of this study.}

\item{num_question_Est}{How many questions used in this estimation.}

\item{type_theta}{Type of scaling response noise parameter used in this estimation, specify either "Global", "Individual" or "Hier".}

\item{path}{Full path for working directory.}

\item{parameter1}{First parameter. Specify either "alpha", "sigma" or "lambda".}

\item{parameter2}{Second parameter. Specify another parameter in "alpha", "sigma" or "lambda".}
}
\value{
Return the distribution plot for all individual level correlation coefficients between two parameters specified.
}
\description{
For each subject, the function firstly computes the correlation coefficient between posterior sampling drawers of two parameters set by \code{parameter1} and \code{parameter2}. Then it draws the distribution of all correlation coefficients across all subjects. 
Only run this function when stanfit object for study \code{project_name} is saved under directory \code{path} after estimation.
}
\examples{
Risk_plot_cor(project_name = 'test', num_question_Est = 12, type_theta = 'Hier', path = '/Users/ap/Desktop', 
parameter1 = 'alpha', parameter2 = 'sigma')
}
