% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_prepare_Stan.R
\name{Risk_prepare_Stan}
\alias{Risk_prepare_Stan}
\title{Prepare Data for Stan Risk Estimation}
\usage{
Risk_prepare_Stan(all_Stan_data, num_question_Est, subjectNumber)
}
\arguments{
\item{all_Stan_data}{Data needed to be processed.}

\item{num_question_Est}{How many questions used in the estimation.}

\item{subjectNumber}{Number of subjects in this survey.}
}
\value{
Several objects will be returned into global enviornment:
\itemize{
  \item ss_amnt, 
  \item 
  \item moddat, a list object containing all data needed in estimation model.
}
}
\description{
Create an R list object containing all necessary data to run estimation for DEEP Risk.
Called by function \code{\link{Stan_Risk_Estimation}}.
}
\examples{
Risk_prepare_Stan(all_Stan_data = all_Stan_data, num_question_Est = 12, subjectNumber = 200)
}
