% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_resetxy.R
\name{Risk_resetxy}
\alias{Risk_resetxy}
\title{Title}
\usage{
Risk_resetxy(g1g2)
}
\arguments{
\item{g1g2}{One row with contexts for two options in one Risk question.}
}
\value{
Arranged data.
}
\description{
Prepare data and make sure the values of x and y fall into the two situations of CPT model indicated in original paper: either \eqn{x>y>0|x<y<0}, or \eqn{x < 0 < y}.
Called by function \code{\link{Stan_Risk_Estimation}}.
}
\examples{
Risk_resetxy(g1g2)
}
