% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_save_stantocsv.R
\name{Risk_save_stantocsv}
\alias{Risk_save_stantocsv}
\title{Save Risk Preferences Estimates to csv Files}
\usage{
Risk_save_stantocsv(project_name, num_question_Est, type_theta, path)
}
\arguments{
\item{project_name}{The name of this study.}

\item{num_question_Est}{How many questions you want to use in estimation.}

\item{type_theta}{Type of scaling response noise parameter used in estimation, specify either "Global", "Individual" or "Hier".}

\item{path}{Full path for working directory.}
}
\value{
Two csv files will be saved under directory indicated by \code{path}:
\itemize{
  \item "StanEstimates_Risk_parameters_\{\code{project_name}\}_Est\{\code{type_theta}\}\{\code{num_question_Est}\}questions.csv" contains all estimates of individual \eqn{\alpha} (distortion of probability parameter), \eqn{\sigma} (curvature of value function) and \eqn{\lambda} (loss aversion parameter) respectively as in three columns. The number of rows is the number of subjects in the survey. 
  \item "StanEstimates_Risk_theta_\{\code{project_name}\}_Est\{\code{type_theta}\}\{\code{num_question_Est}\}questions.csv" contains estimates of either global or individual scaling response error parameter \eqn{\theta}.
}
}
\description{
Save posterior point estimates for risk preferences from stanfit object to local csv files. 
Only run this function when stanfit object for study \code{project_name} is saved under directory \code{path} after estimation.
}
\examples{
Risk_save_stantocsv(project_name = 'test', num_question_Est = 12, type_theta = 'Hier', path = path)
}
\references{
Toubia, O., Johnson, E., Evgeniou, T., & Delquié, P. (2013). Dynamic experiments for 
estimating preferences: An adaptive method of eliciting time and risk parameters. 
Management Science, 59(3), 613-640.
\url{https://pubsonline.informs.org/doi/abs/10.1287/mnsc.1120.1570}
}
