% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stan_Time_Estimation.R
\name{Stan_Time_Estimation}
\alias{Stan_Time_Estimation}
\title{Stan Estimation Function for DEEP Time}
\usage{
Stan_Time_Estimation(project_name, num_question_Est, num_question,
  type_theta, path, save_out = T, chains = 3, iter = 1000,
  thin = 3, adapt_delta = 0.9, max_treedepth = 12, stepsize = 1)
}
\arguments{
\item{project_name}{The name of this study.}

\item{num_question_Est}{How many questions you want to use in estimation.}

\item{num_question}{How many questions are asked for each subject.}

\item{type_theta}{Type of scaling response noise parameter used in estimation, specify either "Global", "Individual" or "Hier".}

\item{path}{Full working path.}

\item{save_out}{Whether save the stanfit object as a rdata file. The default is "TRUE".}

\item{chains}{A positive integer specifying the number of Markov chains. The default is 3.}

\item{iter}{A positive integer specifying the number of iterations for each chain (including warmup). The default is 1000.}

\item{thin}{A positive integer specifying the period for saving samples. The default is 3. For details, refer to help page for function \code{\link[rstan]{stan}}.}

\item{adapt_delta}{A double value between 0 and 1 controlling the accept probability in sampling. The default is 0.9. For details, refer to help page for function \code{\link[rstan]{stan}}.}

\item{max_treedepth}{A positive integer specifying how deep in tree exploration. The default is 12. For details, refer to help page for function \code{\link[rstan]{stan}}.}

\item{stepsize}{A double and positive value controlling sampler's behavior. The default is 1. For details, refer to help page for function \code{\link[rstan]{stan}}.}
}
\value{
Return a large stanfit object called "hier_time" if \code{save_out=FALSE}. Otherwise, return nothing but save the stanfit object into a local RData file named 
"Stan_Time_\{\code{project_name}\}_Est\{\code{type_theta}\}\{\code{num_question_Est}\}questions.RData" under directory \code{path}.
}
\description{
The main function to do estimation on DEEP Time data. With
}
\examples{
Stan_Time_Estimation(project_name = 'Test', num_question_Est = 12, num_question = 12, 
type_theta = 'Hier', path = '/Users/ap/Desktop')
}
\references{
Toubia, O., Johnson, E., Evgeniou, T., & Delquié, P. (2013). Dynamic experiments for 
estimating preferences: An adaptive method of eliciting time and risk parameters. 
Management Science, 59(3), 613-640.
\url{https://pubsonline.informs.org/doi/abs/10.1287/mnsc.1120.1570}
}
