% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_data_prepare.R
\name{Time_data_prepare}
\alias{Time_data_prepare}
\title{Survey Data Read, Clean and Denoise}
\usage{
Time_data_prepare(project_name, path, num_question, clean_time = T,
  clean_side = T, each_seconds = 3)
}
\arguments{
\item{project_name}{The name of this study.}

\item{path}{Full path for working directory.}

\item{num_question}{How many questions are asked for each subject in this suvery.}

\item{clean_time}{Logical value. Default is TRUE means you want to denoise the original data by dropping those}

\item{clean_side}{}

\item{each_seconds}{}
}
\value{
Three csv files will be saved under directory indicated by \code{path}:
\itemize{
  \item "\{\code{project_name}\}_Time_options_chosen.csv" contains all estimates of individual \eqn{\beta} (present-bias parameter), \eqn{r} (daily discount rate) and yearly discount factor \eqn{\delta} respectively as in three columns. The number of rows is the number of subjects in the survey. 
  \item "\{\code{project_name}\}_Time_options_notchosen.csv" contains estimates of either global or individual scaling response error parameter \eqn{\theta}.
  \item "\{\code{project_name}\}_Time_serials.csv" contains serial number for each subject as identification.
}
}
\description{
Read Survey
}
\examples{
Time_data_prepare(project_name = 'test', path = '/Users/ap/Desktop', num_question = 12)
}
