% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_dist_estimates.R
\name{Time_dist_estimates}
\alias{Time_dist_estimates}
\title{Plot the Distribution of Estimates for Time Preference Parameters.}
\usage{
Time_dist_estimates(project_name, num_question_Est, type_theta, path)
}
\arguments{
\item{project_name}{The name of this study.}

\item{num_question_Est}{How many questions used in the estimation.}

\item{type_theta}{Type of scaling response noise parameter used in estimation, specify either "Global", "Individual" or "Hier".}

\item{path}{Full path for working directory.}
}
\value{
Return the distribution plot for all parameter estimates for all subjects.
}
\description{
Plot the distributions of all parameter estimates across all subjects.
Only run this function when stanfit object for study \code{project_name} is saved under directory \code{path} after estimation.
}
\examples{
Time_dist_estimates(project_name = 'test', num_question_Est = 12, type_theta = 'Hier', path = '/Users/ap/Desktop')
}
