% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_index_Questions.R
\name{Time_index_Questions}
\alias{Time_index_Questions}
\title{Index and Arrange Time Questions}
\usage{
Time_index_Questions(stan_data, subjectNumber, num_question)
}
\arguments{
\item{stan_data}{Data needed to be processed.}

\item{subjectNumber}{Number of subjects in this survey.}

\item{num_question}{How many questions are asked for each subject in this survey.}
}
\value{
Return the indexed and arranged data.
}
\description{
Index and arrange questions so that smaller-sooner and larger-later options would be arranged as expected. 
Called by function \code{\link{Stan_Time_Estimation}}.
}
\examples{
Time_Index_Questions(stan_data = stan_data, subjectNumber = 200, num_question = 20)
}
