% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_save_stantocsv.R
\name{Time_save_stantocsv}
\alias{Time_save_stantocsv}
\title{Save Time Preferences Estimates to csv Files}
\usage{
Time_save_stantocsv(project_name, num_question_Est, type_theta, path)
}
\arguments{
\item{project_name}{The name of this study.}

\item{num_question_Est}{How many questions you want to use in estimation.}

\item{type_theta}{Type of scaling response noise parameter used in estimation, specify either "Global", "Individual" or "Hier".}

\item{path}{Full working path.}
}
\value{
Two csv files will be saved under directory indicated by \code{path}:
\itemize{
  \item "StanEstimates_Time_parameters_\{\code{project_name}\}_Est\{\code{type_theta}\}\{\code{num_question_Est}\}questions.csv" contains all estimates of individual \eqn{\beta} (present-bias parameter), \eqn{r} (daily discount rate) and yearly discount factor \eqn{\delta} respectively as in three columns. The number of rows is the number of subjects in the survey. 
  \item "StanEstimates_Time_theta_\{\code{project_name}\}_Est\{\code{type_theta}\}\{\code{num_question_Est}\}questions.csv" contains estimates of either global or individual scaling response error parameter \eqn{\theta}.
}
}
\description{
Save posterior point estimates for time preferences from stanfit object to csv files.
}
\examples{
Time_save_stantocsv(project_name = 'test', num_question_Est = 12, type_theta = 'Hier', path = path)
}
