% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEPEst-package.R
\docType{package}
\name{DEEPEst-package}
\alias{DEEPEst-package}
\alias{DEEPEst}
\title{The 'DEEPEst' package.}
\description{
Dynamic Experiments for Estimating Preferences (DEEP) is a novel methodology 
to elicit individuals' risk and time preferences by dynamically optimizing the sequences of 
questions presented to each subject, while leveraging information about the distribution of 
the parameters across individuals (heterogeneity) and modeling response error explicitly. 
This package provides functions to execute hierarchical Bayesian estimation for DEEP with 'Stan' <https://mc-stan.org> 
and posterior analysis on estimates of both time and risk preferences parameters.
}
\references{
Toubia, O., Johnson, E., Evgeniou, T., & Delquié, P. (2013). Dynamic experiments for estimating preferences: An adaptive method of eliciting time and risk parameters. 
Management Science, 59(3), 613-640. \url{https://pubsonline.informs.org/doi/abs/10.1287/mnsc.1120.1570}
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
