% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_prepare_Stan.R
\name{Risk_prepare_Stan}
\alias{Risk_prepare_Stan}
\title{Prepare Data for Stan Risk Estimation}
\usage{
Risk_prepare_Stan(all_Stan_data, num_question_Est, subjectNumber)
}
\arguments{
\item{all_Stan_data}{Data needed to be processed.}

\item{num_question_Est}{How many questions are going to be used in this estimation.}

\item{subjectNumber}{Number of subjects in this survey.}
}
\value{
Several objects will be returned into global enviornment:
\itemize{
  \item x1: a dataframe contains amounts of first rewards in the left options. 
  \item p1: a dataframe contains the probabilities receiving \code{x1}.
  \item y1: a dataframe contains amounts of second rewards in the left options.
  \item q1: a dataframe contains the probabilities receiving \code{y1}.
  \item x2: a dataframe contains amounts of first rewards in the right options.
  \item p2: a dataframe contains the probabilities receiving \code{x2}.
  \item y2: a dataframe contains amounts of second rewards in the right options.
  \item q2: a dataframe contains the probabilities receiving \code{y2}.
  \item choices: a dataframe contains all choices made by subjects with "0" means left option is chosen and "1" means right option is chosen.
  \item moddat: a list object contains all these data needed in estimation model.
}
}
\description{
After reading, processing and denoising necessary data. This function will create an R list object containing all data to run estimation for DEEP Risk.
Called by function \code{\link{Stan_Risk_Estimation}}.
}
\examples{
Risk_prepare_Stan(all_Stan_data = all_Stan_data, num_question_Est = 12, subjectNumber = 200)
}
