% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_data_prepare.R
\name{Time_data_prepare}
\alias{Time_data_prepare}
\title{Read, Clean and Denoise Time Survey Data}
\usage{
Time_data_prepare(project_name, path, num_question, clean_time = T,
  clean_side = T, each_seconds = 3)
}
\arguments{
\item{project_name}{The name of this study.}

\item{path}{Full path for working directory.}

\item{num_question}{How many questions are asked for each subject in this suvery.}

\item{clean_time}{Logical value. Default is \code{TRUE} means you want to denoise the original data by dropping all responses which are completed within \code{num_question} times \code{each_seconds} seconds.}

\item{clean_side}{Logical value. Default is \code{TRUE} means you want to denoise the original data by dropping all responses which only contain one-side options, namely, only left or right options chosen.}

\item{each_seconds}{How many seconds spent on each question on average so that the observation would be considered as a reasonable response.}
}
\value{
Three csv files will be saved under directory indicated by \code{path} and used in estimation function \code{\link{Stan_Time_Estimation}}:
\itemize{
  \item "\{\code{project_name}\}_Time_options_chosen.csv" contains amounts of rewards (first column) and delayed days (second column) of all chosen options. The number of rows is \code{num_question} times \code{num_subjects}, where \code{num_subjects} is the number of subjects in this survey after denoising.
  \item "\{\code{project_name}\}_Time_options_notchosen.csv" is same as the above one except this contains data of unchosen options among all questions.
  \item "\{\code{project_name}\}_Time_serials.csv" contains serial number for each subject as identification.
}
}
\description{
Read and denoise the original suvery output data. Only complete rows would be reserved. In addition, this function automatically drops those "unengaged" responses (noisy ones with only random choices, not involving utility consideration).
Once you obtain the original output csv file from DEEP survey under directory \code{path}, rename it to be "DEEP_\{\code{model}\}_surveydata_\{\code{project_name}\}.csv", where "\code{model}" is either "Time" or "Risk".
}
\examples{
Time_data_prepare(project_name = 'test', path = '/Users/ap/Desktop', num_question = 12)
}
